+++ app/src/main/java/com/boosters/promise/data/invite/FriendRepository.kt
@@ -1,0 +1,9 @@
+package com.boosters.promise.data.invite
+
+import com.boosters.promise.ui.model.UserState
+
+interface FriendRepository {
+
+    fun getFriends(): List<UserState>
+
+}
