--- app/src/main/java/com/boosters/promise/invite/FriendRepository.kt
@@ -1,5 +1,0 @@
-package com.boosters.promise.invite
-
-interface FriendRepository {
-    fun getFriends(): List<User>
-}
