--- app/src/main/java/com/boosters/promise/invite/InviteDiffItemCallback.kt
@@ -1,13 +1,0 @@
-package com.boosters.promise.invite
-
-import androidx.recyclerview.widget.DiffUtil
-
-class InviteDiffItemCallback : DiffUtil.ItemCallback<User>() {
-    override fun areItemsTheSame(oldItem: User, newItem: User): Boolean {
-        return oldItem.userCode == newItem.userCode
-    }
-
-    override fun areContentsTheSame(oldItem: User, newItem: User): Boolean {
-        return oldItem == newItem
-    }
-}
