--- app/src/main/java/com/boosters/promise/invite/User.kt
@@ -1,6 +1,0 @@
-package com.boosters.promise.invite
-
-data class User(
-    val userName: String,
-    val userCode: String
-)
