+++ app/src/main/java/com/boosters/promise/ui/model/UserState.kt
@@ -1,0 +1,6 @@
+package com.boosters.promise.ui.model
+
+data class UserState(
+    val userName: String,
+    val userCode: String
+)
