+++ app/src/main/java/com/boosters/promise/util/BindingAdapter.kt
@@ -1,4 +1,4 @@
-package com.boosters.promise.invite
+package com.boosters.promise.util
 
 import androidx.databinding.BindingAdapter
 import androidx.recyclerview.widget.ListAdapter
@@ -7,6 +7,7 @@ import androidx.recyclerview.widget.RecyclerView
 @Suppress("UNCHECKED_CAST")
 @BindingAdapter("items")
 fun <T> RecyclerView.setItems(items: List<T>?) {
+
     if (items.isNullOrEmpty()) {
         return
     }
@@ -14,4 +15,5 @@ fun <T> RecyclerView.setItems(items: List<T>?) {
     (adapter as ListAdapter<T, RecyclerView.ViewHolder>).submitList(items) {
         scrollToPosition(0)
     }
+
 }
