+++ app/src/main/res/layout/activity_invite.xml
@@ -4,7 +4,7 @@
     <data>
         <variable
             name="viewModel"
-            type="com.boosters.promise.invite.InviteViewModel" />
+            type="com.boosters.promise.ui.invite.InviteViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
@@ -12,7 +12,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".invite.InviteActivity">
+        tools:context=".ui.invite.InviteActivity">
 
         <androidx.constraintlayout.widget.Guideline
             android:id="@+id/member_left_guideline"
