+++ app/src/main/res/layout/item_invite_member_info.xml
@@ -5,7 +5,7 @@
 
         <variable
             name="user"
-            type="com.boosters.promise.invite.User" />
+            type="com.boosters.promise.ui.model.UserState" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
